+++ src/dataframe/dataframe.jl
@@ -1807,14 +1807,14 @@ function _replace_columns!(df::DataFrame, newdf::DataFrame)
     return df
 end
 
-expandgrid(; kwargs...) = isempty(kwargs) ? DataFrame() : expandgrid(kwargs...)
+allcombinations(; kwargs...) = isempty(kwargs) ? DataFrame() : allcombinations(kwargs...)
 
-expandgrid(pairs::Pair{<:AbstractString, <:Any}...) =
-    expandgrid((Symbol(k) => v for (k,v) in pairs)...)
+allcombinations(pairs::Pair{<:AbstractString, <:Any}...) =
+    allcombinations((Symbol(k) => v for (k, v) in pairs)...)
 
 """
-    expandgrid(pairs::Pair...)
-    expandgrid(; kwargs...)
+    allcombinations(pairs::Pair...)
+    allcombinations(; kwargs...)
 
 Create a `DataFrame` from all combinations of values in passed arguments.
 
@@ -1830,7 +1830,7 @@ or a `0`-dimensional `AbstractArray` are unwrapped and treated as having length 
 # Examples
 
 ```jldoctest
-julia> expandgrid(a=1:2, b='a':'c')
+julia> allcombinations(a=1:2, b='a':'c')
 6×2 DataFrame
  Row │ a      b
      │ Int64  Char
@@ -1842,7 +1842,7 @@ julia> expandgrid(a=1:2, b='a':'c')
    5 │     1  c
    6 │     2  c
 
-julia> expandgrid("a" => 1:2, "b" => 'a':'c', "c" => "const")
+julia> allcombinations("a" => 1:2, "b" => 'a':'c', "c" => "const")
 6×3 DataFrame
  Row │ a      b     c
      │ Int64  Char  String
@@ -1855,10 +1855,10 @@ julia> expandgrid("a" => 1:2, "b" => 'a':'c', "c" => "const")
    6 │     2  c     const
 ```
 """
-function expandgrid(pairs::Pair{Symbol, <:Any}...)
+function allcombinations(pairs::Pair{Symbol, <:Any}...)
     colnames = first.(pairs)
     if !allunique(colnames)
-        throw(ArgumentError("All column names passed to expandgrid must be unique"))
+        throw(ArgumentError("All column names passed to allcombinations must be unique"))
     end
     colvalues = map(pairs) do p
         v = last(p)